Added some options to clu_cert_setup.c
Fix detail

Fix extensions sect

Fix hash opt

fix detail

Fix details & added unit test

+++ tests/x509/x509-req-test.sh
@@ -143,6 +143,36 @@ if [ $? != 0 ]; then
 fi
 rm -f tmp.csr.pem
 rm -f tmp.csr.der
+
+# test passing csr file for x509
+run_fail "x509 -in tmp.csr -days 3650 -out tmp.crt"
+run_fail "x509 -req -in tmp.csr -days 3650 -out tmp.crt"
+run_success "x509 -req -in tmp.csr -days 3650 -signkey ./certs/server-key.pem -out tmp.crt"
+rm -f tmp.crt
+
+
+#testing hash for x509
+run_success "x509 -req -in tmp.csr -days 3650 -md sha -signkey ./certs/server-key.pem -out tmp.crt"
+rm -f tmp.crt
+run_success "x509 -req -in tmp.csr -days 3650 -md sha224 -signkey ./certs/server-key.pem -out tmp.crt"
+rm -f tmp.crt
+run_success "x509 -req -in tmp.csr -days 3650 -md sha256 -signkey ./certs/server-key.pem -out tmp.crt"
+rm -f tmp.crt
+run_success "x509 -req -in tmp.csr -days 3650 -md sha384 -signkey ./certs/server-key.pem -out tmp.crt"
+rm -f tmp.crt
+run_success "x509 -req -in tmp.csr -days 3650 -md sha512 -signkey ./certs/server-key.pem -out tmp.crt"
+rm -f tmp.crt
+
+#testing extensions for x509
+run_success "x509 -req -in tmp.csr -days 3650 -extfile ./test.conf -extensions v3_alt_ca -signkey ./certs/server-key.pem -out tmp.crt"
+run_success "x509 -in tmp.crt -text -noout"
+echo "$RESULT" | grep "CA:TRUE"
+if [ $? != 0 ]; then
+    echo "was expecting alt extensions to have CA set"
+    exit 99
+fi
+
+rm -f tmp.crt
 rm -f tmp.csr
 
 run_success "req -new -key ./certs/server-key.pem -config ./test.conf -x509 -out tmp.cert"
